Implement GSP-73 Organization rename

+++ iterator.go
@@ -58,5 +58,5 @@ type partPageStatus struct {
 }
 
 func (i *partPageStatus) ContinuationToken() string {
-	return strconv.FormatInt(int64(i.partNumberMarker), 10)
+	return strconv.FormatInt(i.partNumberMarker, 10)
 }
